Updated the travel distance to be an int so it does not retur number.123(example)

+++ CarPooling.Data/Models/Travel.cs
@@ -27,7 +27,7 @@ namespace CarPooling.Data.Models
         public DateTime? DepartureTime { get; set; }
         public double EstimatedTravelDuration { get; set; }
         public DateTime? ArrivalTime { get; set; }
-        public double TravelDistance { get; set; }
+        public int TravelDistance { get; set; }
 
         //  public bool? IsCompleted { get; set; }
         public bool? IsCompleted { get; set; }
